arreglo de boton }

+++ grpc-server/src/main/java/com/grpc/grpc_server/repositories/kafka/DonationRequestRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.grpc.grpc_server.entities.kafka.OperationDonation;
 
-public interface DonationRequestRepository extends JpaRepository<OperationDonation, Long> {
+public interface DonationRequestRepository extends JpaRepository<OperationDonation, java.lang.Integer> {
 }
